requirements.txt
nba_api
django-ninja